requirements.txt
streamlit==1.28.1
openai==1.3.7
pinecone-client==2.2.4
langchain==0.0.335
langchain-openai==0.0.2
langchain-pinecone==0.0.1
python-dotenv==1.0.0
watchdog==3.0.0
pandas==2.1.3
markdown==3.5.1
